Change FlightState PK to byte, update seeding and cleanup
Changed FlightState primary key from short to byte and updated data seeding to use integer Ids. Removed unused methods from FlightRepository and performed minor formatting and import cleanups in context files. No logic changes to IsOriginAirportActiveAsync.

+++ Core/AeroVeloz.Domain/Entities/Flights/FlightState.cs
@@ -2,7 +2,7 @@
 
 namespace AeroVeloz.Domain.Entities.Flights;
 
-public partial class FlightState : BEntity<short>
+public partial class FlightState : BEntity<byte>
 {
     public string? code { get; init; }
     public string? StateName { get; init; }
